Small refactor over the charm_state.py
