+++ tests/unit/conftest.py
@@ -12,6 +12,7 @@ from github_runner_manager.configuration.github import GitHubOrg
 
 import charm_state
 import utilities
+from models import FlavorLabel, OpenStackCloudsYAML
 
 
 @pytest.fixture(name="exec_command")
@@ -132,7 +133,7 @@ def complete_charm_state_fixture():
         charm_config=charm_state.CharmConfig(
             dockerhub_mirror="https://docker.example.com",
             labels=("label1", "label2"),
-            openstack_clouds_yaml=charm_state.OpenStackCloudsYAML(
+            openstack_clouds_yaml=OpenStackCloudsYAML(
                 clouds={
                     "microstack": {
                         "auth": {
@@ -161,7 +162,7 @@ def complete_charm_state_fixture():
             base_virtual_machines=1,
             max_total_virtual_machines=2,
             flavor_label_combinations=[
-                charm_state.FlavorLabel(
+                FlavorLabel(
                     flavor="flavor",
                     label="flavorlabel",
                 )
